Fix Linter errors on all classes
Rename files to use snake casing and also refactor code to remove useless
 variable assignments
