+++ student.rb
@@ -1,12 +1,12 @@
-require "./Person.rb"
+require './Person'
 
 class Student < Person
   def initialize(classroom)
-    super(name, age, parent_permission = true)
+    super(name, age, parent_permission)
     @classroom = classroom
   end
 
   def play_hooky
-    "¯\(ツ)/¯"
+    '¯(ツ)/¯'
   end
 end
