+++ teacher.rb
@@ -1,12 +1,12 @@
-require "./Person.rb"
+require './Person'
 
 class Teacher < Person
   def initialize(specialization)
-    super(name, age, parent_permission = true)
+    super(name, age, parent_permission)
     @specialization = specialization
   end
 
   def can_use_services?
-   true
+    true
   end
 end
